Support spying on unbound functions on classes.
spy_on() can now be called on an unbound function on a class (such as
ClassName.my_function), which is useful when the calling code doesn't
have a particular instance it's working on, but needs to spy on an
instance that will be created in the future.

This works just like spying on any other function.

Reviewed at https://reviews.reviewboard.org/r/5910/