Add a test runner for the kgb test suite.
This is a simple test runner using nose that can run the entire test
suite. It will run Python 3 tests out of a tests/py3/ directory when
running on Python 3, and will ignore it when running on Python 2.

The name of the tests script is the same as for our other projects,
ensuring consistency.