Add initial support for Python 3.x.
This updates kgb to work with both Python 2.x and 3.x. There were a
handful of changes needed to support 3.x, but they mostly amounted to
changes in attributes for functions and classes, the addition of one
more parameter for the `CodeType` constructor, and a switch to Unicode
strings.

This change does not support keyword-only arguments, which will require
additional work.

Testing Done:
Ran unit tests on Python 2.7 and 3.5.

Reviewed at https://reviews.reviewboard.org/r/9330/