Fix unit tests on Python 2.6.
The test suite was making use of testing functionality present only on
Python 2.7 and higher, breaking unit tests on older versions of Python.
To fix this, we now make use of the `unittest2` package, which provides
backports of the testing functionality. This is only installed on older
versions of Python.

Testing Done:
Unit tests pass on Python 2.6, 2.7, 3.4, 3.5, and 3.6.

Reviewed at https://reviews.reviewboard.org/r/9519/