Modernize the documentation and switch to ReST.
The documentation in the README was a bit out of date when it comes to
kgb 2.0, talking about the necessity of using the `.spy` attribute for
public functions and not including all supported Python versions. It was
also in Markdown, meaning it wasn't helpful at all for PyPI.

This change moves the README to ReST, updates the contents to be more
correct with kgb 2.0, and attaches it to the package metadata.

Testing Done:
Rendered the README and read through it. Didn't spot any spelling errors
or other issues.

Reviewed at https://reviews.reviewboard.org/r/9593/