Fix spying on a classmethod defined on a parent class.
Spying on a classmethod of a class would cause spy leaks if the
classmethod was actually defined on a parent. All subclasses of the
parent would carry the same spy, causing false results and breakages.

We now check if the classmethod being spied on now exists in a parent of
the class, and if so, we clone it. This prevents spy leaks.

Unit tests were added to check this condition.

Testing Done:
Unit tests pass on all supported versions of Python.

Tested against issues hit during development. The spies succeeded and
no longer leaked.

Reviewed at https://reviews.reviewboard.org/r/10072/