Add assertSpyNotCalledWith().
This fills a gap in the assertion methods offered by KGB, allowing a
caller to assert that a particular call was never made. This is the
negation of `assertSpyCalledWith()`, and accepts a function, a spy, or a
call.

Testing Done:
Unit tests pass for all supported versions of Python.

Used this in a unit test successfully.

Reviewed at https://reviews.reviewboard.org/r/11159/