Convert the NEWS file to ReStructuredText.
The NEWS file felt pretty crufty, since it was all just plain text.
As a step toward eventually having some solid documentation with a set
of release notes as part of it, this change rewrites this file to be in
ReST format. This helps with the presentation, allows us to link to
content, and just makes for a nicer reading experience.

Testing Done:
Viewed the content in a ReST renderer. Checked for any rendering issues.

Reviewed at https://reviews.reviewboard.org/r/11160/