Allow spy operations to nest spy operations.
Spy operations that take list of expected calls can now define an `op`
key in a call for a nested spy operation. This allows them to take
advantage of useful ops like `SpyOpReturn` or `SpyOpRaise`. It also
allows for more complex rulesets that nest `SpyOpMatchInOrder` within a
`SpyOpMatchAny`, which is useful when there's a lot of calls to a
utility function with different arguments.

This will be the major feature for 6.1.

Testing Done:
Unit tests pass for all supported versions of Python.

Reviewed at https://reviews.reviewboard.org/r/11797/