Improve the error for unexpected function call spies in operations.
In match-based spy operations, exceptions are raised when a call is made
to a spy that either exceeds or doesn't match the list of expected
calls.

To help with debugging these situations, the exception now includes the
call that triggered the exception.

This uncovered an issue in `SpyCall.__repr__` where a `)` was missing
near the end of the string.

Testing Done:
Unit tests pass on Python 2 and 3.

Reviewed at https://reviews.reviewboard.org/r/11798/