Add new SpyOpReturnInOrder and SpyOpRaiseInOrder spy operations.
`SpyOpReturnInOrder` is a mix between `SpyOpReturn` and
`SpyOpMatchInOrder`. It takes a list of values, and will return a value
for each call made, in the order provided. If a call once the list is
exhausted, an exception will be raised.

`SpyOpRaiseInOrder` is the `SpyOpRaise` equivalent. It will raise
exceptions in order.

Testing Done:
All unit tests pass in Python 2 and 3.

Reviewed at https://reviews.reviewboard.org/r/11799/