Add release notes for kgb 6.1.
This is a feature release that focuses on some new spy capabilities:

1. Nested operations inside `SpyOpMatchInOrder`/`SpyOpMatchAny`
2. `SpyOpReturnInOrder`
3. `SpyOpRaiseInOrder`

This will be a 6.1 release, in an effort to finally move to semantic
versioning. From here on out, 6.x will be the target for features, until
we need to break compatibility.

Testing Done:
Viewed the rendered release notes. Checked for spelling errors.

Reviewed at https://reviews.reviewboard.org/r/11800/