Fix unit tests on Python 3.10 and update listed compatibility.
This adds support for the newly-released Python 3.10. kgb itself had no
problems with this release that we could tell, but unit tests did fail
due to `nose` not being compatible with 3.10 (it uses
`collections.Callable`, which no longer exists).

We now use `nose3` on Python 3. This is a drop-in replacement for `nose`
with modern Python 3 fixes.

Along with this, we're dropping Python 2.5, 2.6, 3.4, and 3.5 support.
We haven't been able to really test these in a while. Since technically
this is a breaking change, and we've moved to semantic versioning, we're
bumping kgb up to version 7.

Testing Done:
Unit tests pass on all supported versions of Python (2.7 and 3.6-3.10).

Reviewed at https://reviews.reviewboard.org/r/11843/