Update the package version generation and Python compatibility.
We've relied on setuptools to convert our `alpha`/`beta` tags to
`a`/`b`. However, this generates an annoying warning. To avoid this, we
now handle normalization as part of string generation.

The package also now defines a `python_requires`, which specifies all
versions of Python that are supported. This will help `pip` determine
which versions of kgb are eligible for installation in the future.

Testing Done:
Built the package. Didn't see any warnings.

Inspected the metadata for the new Python requirements.

Reviewed at https://reviews.reviewboard.org/r/11922/